adding higher order visitor decorator functions. Starting to reorganize visitors into a separated folder
